Support BF16 for Engine/Tune (#546)
* update the pybind to support bf16 cast

* support bf16 for matmul/innerproduct and model/common

* strength padding_sequence and change uint16 to bf16

+++ engine/compile/graph_utils.py
@@ -29,7 +29,7 @@ DTYPES_DICT = {"float16": "fp16",
                "int32": "s32",
                "int64": "s64",
                "uint8": "u8",
-               "uint16": "u16",
+               "uint16": "bf16",
                }
